User and spot table associations and validations complete

+++ backend/db/migrations/20231128005805-create-spot.js
@@ -50,7 +50,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       price: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(2,2),
         allowNull: false
       },
       createdAt: {
